moved more of the logic out of the app.jsx
